Add logging message and move transform function into utils
